app/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudscraper==1.2.64
commonmark==0.9.1
exceptiongroup==1.0.0rc9
filelock==3.8.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
platformdirs==2.5.2
pyaes==1.6.1
pyasn1==0.4.8
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.7.0
python-dotenv==0.21.0
pytube==12.1.0
PyVirtualDisplay==3.0
requests==2.28.1
requests-toolbelt==0.10.0
rich==12.6.0
rsa==4.9
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
telebot==0.0.4
Telethon==1.25.2
tiktok-downloader==0.3.4
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
uritools==4.0.0
urlextract==1.6.0
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
xvfbwrapper==0.2.9
